Rework of fs2, (seems to) pass test_core_online_tree_and_sync_multicore.py
