+++ parsec/core/fs2/exceptions.py
@@ -1,0 +1,2 @@
+class InvalidPath(Exception):
+    pass
